I'm stupid

+++ src/server/globals.h
@@ -28,7 +28,7 @@ typedef struct
 {
 	uint64_t handle;          // Client defined handle to pass back in reply
 	uint64_t from;            // First byte offset of requested block (ie. 4096)
-	volatile uint32_t to;     // Last byte + 1 of requested block (ie. 8192, if request len is 4096, resulting in bytes 4096-8191)
+	volatile uint64_t to;     // Last byte + 1 of requested block (ie. 8192, if request len is 4096, resulting in bytes 4096-8191)
 	dnbd3_client_t * volatile client; // Client to send reply to
 	volatile int status;      // status of this entry: ULR_*
 } dnbd3_queued_request_t;
